+++ lib/adjustable_schema/active_record.rb
@@ -3,6 +3,7 @@
 module AdjustableSchema
 	module ActiveRecord # :nodoc:
 		autoload :Association, 'adjustable_schema/active_record/association'
+		autoload :Builder,     'adjustable_schema/active_record/builder'
 
 		ActiveSupport.on_load :active_record do
 			include Associations