+++ test/lib/active_record/association/naming_test.rb
@@ -13,7 +13,7 @@ describe AdjustableSchema::ActiveRecord::Association::Naming do
 
 	let(:owner)     { Model1 }
 	let(:target)    { Model2 }
-	let(:role)      { AdjustableSchema::Relationship::Role.find_or_create_by! name: role_name }
+	let(:role)      { AdjustableSchema.role! role_name }
 	let(:role_name) { :dummy_role }
 
 	describe '#name' do