+++ test/models/relationship_test.rb
@@ -1,8 +1,5 @@
 # frozen_string_literal: true
 
-# rubocop:disable Layout/MultilineBlockLayout
-# rubocop:disable Layout/BlockEndNewline
-
 require 'test_helper'
 require 'minitest/autorun'
 
@@ -34,66 +31,11 @@ module AdjustableSchema
 			end
 		end
 
-		describe '.seed!' do
-			let(:last_seed) { described_class.last }
-
-			after { last_seed.destroy }
-
-			it 'accepts model names' do
-				described_class.seed! 'Model1', 'Model2'
-
-				_ { last_seed.attributes.symbolize_keys => {
-						source_type: 'Model1',
-						source_id:   nil,
-						target_type: 'Model2',
-						target_id:   nil,
-						role_id:     nil,
-				} }.must_pattern_match
-			end
-
-			it 'accepts a Hash-like syntax' do
-				described_class.seed! Model1 => Model2
-
-				_ { last_seed.attributes.symbolize_keys => {
-						source_type: 'Model1',
-						source_id:   nil,
-						target_type: 'Model2',
-						target_id:   nil,
-						role_id:     nil,
-				} }.must_pattern_match
-			end
-
-			it 'accepts a single model' do
-				described_class.seed! Model1
-
-				_ { last_seed.attributes.symbolize_keys => {
-						source_type: 'Model1',
-						source_id:   nil,
-						target_type: 'Model1',
-						target_id:   nil,
-						role_id:     nil,
-				} }.must_pattern_match
-			end
-
-			it 'accepts roles' do
-				described_class.seed! Model1 => Model2, roles: %w[ dummy_role ]
-
-				_ { last_seed.attributes.symbolize_keys => {
-						source_type: 'Model1',
-						source_id:   nil,
-						target_type: 'Model2',
-						target_id:   nil,
-				} }.must_pattern_match
-
-				_(last_seed.name).must_equal 'dummy_role'
-			end
-		end
-
 		describe '#name=' do
 			subject { described_class.new name: role_name }
 
 			let(:role_name) { :dummy_role }
-			let(:role) { described_class::Role.find_or_create_by! name: role_name }
+			let(:role)      { AdjustableSchema.role! role_name }
 
 			before { role } # should exist